Phase 1 Step 7: extract RecommendationsService, kill cross-server tool imports
RecommendationsService composes the five domain service instances (Prices/Options/Microstructure/Sentiment/Fundamentals) plus OhlcvRepository + YFinanceGateway via constructor injection. Absorbs get_trade_recommendation (19 signals), get_stop_loss_analysis, and get_relative_strength(+history) verbatim. stock_price_server.py shrinks 1730->703 lines; the four tools become thin adapters, replacing the fragile pattern of importing each other's decorated tool functions.
